refactor(init): move init files to package-extension
