+++ cogs5e/initiative/__init__.py
@@ -4,3 +4,11 @@ from .effect import Effect
 from .errors import *
 from .group import CombatantGroup
 from .utils import CombatantType
+
+pass  # don't move my imports, pycharm - the cog has to be imported last
+
+from .cog import InitTracker  # noqa E402
+
+
+def setup(bot):
+    bot.add_cog(InitTracker(bot))
